Add language selection feature
Add language selection feature with UI components for choosing
between English and Arabic languages. Includes widgets for
displaying language options, confirming selection, and navigating
to onboarding view. Also update main.dart to use LanguageView
instead of LanguageScreen.
