+++ lib/Features/Language/Presentaion/View/widget/custom_choose_language.dart
@@ -1,35 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-
 import 'package:egytraveler/core/resources/app_localizations.dart';
 import 'package:egytraveler/core/utils/color_manager.dart';
 import 'package:egytraveler/core/utils/styles.dart';
-
-import '../../../generated/assets.dart';
+import '../../../../../generated/assets.dart';
 
 class CustomChooseLanguage extends StatelessWidget {
-  const CustomChooseLanguage(
-      {required this.title,
-      required this.image,
-      required this.isChoose,
-      super.key});
+  const CustomChooseLanguage({
+    required this.title,
+    required this.image,
+    required this.isChoose,
+    super.key,
+  });
 
-  final String title;
-  final String image;
+  final String title; // Language name
+  final String image; // Image asset path for the language flag
+  final bool isChoose; // Indicates if the language is selected
 
-  final bool isChoose;
   @override
   Widget build(BuildContext context) {
     return Container(
       height: 65,
+      // Container to hold language selection
       decoration: ShapeDecoration(
         color: Colors.white,
         shape: RoundedRectangleBorder(
           side: BorderSide(
               width: 2,
               color: !isChoose
-                  ? ColorManager.kColorIdle
-                  : ColorManager.kColorPrimary),
+                  ? ColorManager
+                      .kColorIdle // Border color for unselected language
+                  : ColorManager
+                      .kColorPrimary), // Border color for selected language
           borderRadius: BorderRadius.circular(20),
         ),
       ),
@@ -38,6 +40,7 @@ class CustomChooseLanguage extends StatelessWidget {
           const SizedBox(
             width: 13,
           ),
+          // Container for displaying language flag
           Container(
             width: 48,
             height: 48,
@@ -50,8 +53,10 @@ class CustomChooseLanguage extends StatelessWidget {
                 side: BorderSide(
                     width: 2,
                     color: !isChoose
-                        ? ColorManager.kColorIdle
-                        : ColorManager.kColorPrimary),
+                        ? ColorManager
+                            .kColorIdle // Border color for unselected language
+                        : ColorManager
+                            .kColorPrimary), // Border color for selected language
                 borderRadius: BorderRadius.circular(500),
               ),
             ),
@@ -59,11 +64,13 @@ class CustomChooseLanguage extends StatelessWidget {
           const SizedBox(
             width: 23,
           ),
+          // Text widget for displaying language title
           Text(
-            title.tr(context),
+            title.tr(context), // Translate language title
             style: Styles.textRegular18,
           ),
           const Spacer(),
+          // Display check mark if language is selected
           if (isChoose)
             SvgPicture.asset(
               Assets.imageCheckCircle,
