+++ lib/Features/Language/Presentaion/View/widget/language_confirm_button.dart
@@ -1,11 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:egytraveler/core/resources/app_localizations.dart';
 import 'package:egytraveler/core/utils/color_manager.dart';
 import 'package:egytraveler/core/utils/styles.dart';
 import 'package:egytraveler/layout/homeLayout/cubit/cubit.dart';
 import 'package:egytraveler/Features/onBoarding/Presentation/view/onboarding_view.dart';
 import 'package:egytraveler/shared/network/local/cache_helper.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
 
 class LanguageConfirmButton extends StatelessWidget {
   const LanguageConfirmButton({
@@ -19,11 +19,13 @@ class LanguageConfirmButton extends StatelessWidget {
   Widget build(BuildContext context) {
     return GestureDetector(
       onTap: () {
+        // Save selected language code to cache based on user selection
         if (cubit.isLanguageEn == true) {
           context.read<HomeCubit>().cachedLanguageCode('en');
         } else {
           context.read<HomeCubit>().cachedLanguageCode('ar');
         }
+        // Save user preference and navigate to onboarding view
         CacheHelper.saveData(
           key: 'SettingsPage',
           value: true,
@@ -38,6 +40,7 @@ class LanguageConfirmButton extends StatelessWidget {
         child: Container(
           width: 200,
           height: 56,
+          // Styled container for confirm button
           decoration: ShapeDecoration(
             color: ColorManager.kColorPrimary,
             shape: RoundedRectangleBorder(
@@ -45,9 +48,10 @@ class LanguageConfirmButton extends StatelessWidget {
             ),
           ),
           child: Center(
+            // Text widget for the button label
             child: Text(
               'Done'.tr(context),
-              style: Styles.textWhiteBold22,
+              style: Styles.textBold20,
             ),
           ),
         ),
